+++ NEWS.md
@@ -23,6 +23,7 @@ wham 1.0.5.9000
 * error reported in [issue 48](https://github.com/timjmiller/wham/issues/48). There is no `opt` element when do.fit = FALSE.
 * `plot_wham_output` now removes any existing plot files in the directories where plots are saved before writing files so that any orphaned files from previous fits are not included in the results and html rendering.
 * `plot_wham_output` now creates `.png` output by default because of 1) issues opening the `.html` in Firefox on Windows, and 2) it was annoying to close dozens of browser tabs during testing. issues [42](https://github.com/timjmiller/wham/issues/42) and [50](https://github.com/timjmiller/wham/issues/50), thanks to @ejosymart and @JDeroba.
+* `set_catch` bug when catch age comp data exist for some years and not others
 
 wham 1.0.5 (2021-08-25)
 =========================